Added function 'get_literal_locations_structure'.
This function returns two lists of lists, the first representing the positive variables and the second representing the negative variables.
Each position of a list represents a variable, and the list that is in that position has the number of the clauses that have the variable in them.

- Example:
Clause {1, -2, 3} in position 7.
	positive_literals_location[1 -1].append(7)
	negative_literals_location[|-2| -1].append(7)
	positive_literals_location[3 -1].append(7)